#68 removed all shell dependency from build script, though it now depends on Zip extension. Folder for includes added with .gitignore to keep in in repository.

+++ build/buildOperaExtension.php
@@ -7,27 +7,86 @@ include "BuildExtension.php";
 
 class BuildOperaExtension extends BuildExtension {
     const EXT_DIR                   = "extensions/opera/dirtyMSP/";
+    const EXT_PRIVATE_BUILD_DIR     = "extensions/opera/.build/";
     const EXT_FILE_NAME_TEMPLATE    = "dirtyMSP-%s.oex";
+    const EXT_RES_JS_FILE           = "includes/payload.js";
+    const EXT_JS_TEMPLATE           = "../payload.template.js";
     const SUPER_VERSION             = 1;
+    const BUILD_ERROR_CODE          = 1;
 
     private $tmpPath                = null;
+
+    private function checkPreRequisites() {
+        if (!class_exists("ZipArchive")) {
+            echo "Please install PHP zip extension to use this build script. http://php.net/manual/en/book.zip.php".PHP_EOL;
+            exit(self::BUILD_ERROR_CODE);
+        }
+    }
+
+    public function cleanUp() {
+        $this->deleteRecursive($this->getPrivateBuildDir());
+    }
 
     public function getExtDir() {
         return $this->getRootDir().self::EXT_DIR;
     }
 
+    public function getPrivateBuildDir() {
+        return $this->getRootDir().self::EXT_PRIVATE_BUILD_DIR;
+    }
+
     public function getExtTmpDir() {
         if (is_null($this->tmpPath)) {
-            $this->tmpPath = trim(shell_exec("mktemp -d -t fb-image-build.XXXXXXXXX"));
+            $this->tmpPath = sprintf($this->getPrivateBuildDir()."%s/src", time());
         }
 
         return $this->tmpPath.DIRECTORY_SEPARATOR;
     }
 
-    public function copyCode() {
-        shell_exec("cp -r {$this->getExtDir()} {$this->getExtTmpDir()}");
+    public function prepareCode() {
+        if (mkdir($this->getExtTmpDir(), 0755, true) === false) {
+            echo "Cannot create temporary folder [{$this->getExtTmpDir()}]. Check permissions.".PHP_EOL;
+            exit(self::BUILD_ERROR_CODE);
+        }
 
-        shell_exec("cp {$this->getSourceJsFilePath()} {$this->getResJsFilePath()}");
+        $this->copyRecursive($this->getExtDir(), $this->getExtTmpDir());
+
+        $js = file_get_contents($this->getSourceJsFilePath());
+        $template = file_get_contents($this->getExtDir().self::EXT_JS_TEMPLATE);
+        file_put_contents($this->getResJsFilePath(), str_replace("//@@payload@@", $js, $template));
+
+        unlink(dirname($this->getResJsFilePath()).DIRECTORY_SEPARATOR.".gitignore");
+    }
+
+    public function copyRecursive($from, $to) {
+        foreach (
+            $iterator = new RecursiveIteratorIterator(
+                new RecursiveDirectoryIterator($from, RecursiveDirectoryIterator::SKIP_DOTS),
+                RecursiveIteratorIterator::SELF_FIRST
+            ) as $item
+        ) {
+            if ($item->isDir()) {
+                mkdir($to.DIRECTORY_SEPARATOR.$iterator->getSubPathname());
+            } else {
+                copy($item, $to.DIRECTORY_SEPARATOR.$iterator->getSubPathname());
+            }
+        }
+    }
+
+    public function deleteRecursive($path) {
+        foreach (
+            $iterator = new RecursiveIteratorIterator(
+                new RecursiveDirectoryIterator($path, RecursiveDirectoryIterator::SKIP_DOTS),
+                RecursiveIteratorIterator::CHILD_FIRST
+            ) as $item
+        ) {
+            if ($item->isFile()) {
+                unlink($item->getPathName());
+            } else {
+                rmdir($item->getPathName());
+            }
+        }
+        rmdir($path);
     }
 
     private function getExtFileName() {
@@ -39,11 +98,11 @@ class BuildOperaExtension extends BuildExtension {
     }
 
     public function getExtFilePath() {
-        return $this->getExtTmpDir().$this->getExtFileName();
+        return $this->getExtTmpDir()."..".DIRECTORY_SEPARATOR.$this->getExtFileName();
     }
 
     public function getResJsFilePath() {
-        return $this->getExtTmpDir()."includes/payload.js";
+        return $this->getExtTmpDir().self::EXT_RES_JS_FILE;
     }
 
     public function getManifestFilePath() {
@@ -51,7 +110,26 @@ class BuildOperaExtension extends BuildExtension {
     }
 
     public function packExtension() {
-        shell_exec("cd {$this->getExtTmpDir()}; zip -r -9 {$this->getExtFilePath()} *");
+        $zip = new ZipArchive();
+        if ($zip->open($this->getExtFilePath(), ZIPARCHIVE::CREATE) === false) {
+            echo "Cannot create zip archive at [{$this->getExtFileName()}]".PHP_EOL;
+            die(self::BUILD_ERROR_CODE);
+        }
+
+        foreach (
+            $iterator = new RecursiveIteratorIterator(
+                new RecursiveDirectoryIterator($this->getExtTmpDir(), RecursiveDirectoryIterator::SKIP_DOTS),
+                RecursiveIteratorIterator::CHILD_FIRST
+            ) as $item
+        ) {
+            $itemPath = str_replace($this->getExtTmpDir(), "", $item->getPathName());
+            if (!$item->isDir()) {
+                $zip->addFile($item->getPathName(), $itemPath);
+            }
+            echo "Adding into package: {$item->getPathName()}".PHP_EOL;
+        }
+
+        $zip->close();
     }
 
     public function populateConfig() {
@@ -61,14 +139,19 @@ class BuildOperaExtension extends BuildExtension {
     }
 
     public function moveProductToResultFolder() {
-        unlink($this->getResBundleFilePath());
+        if (file_exists($this->getResBundleFilePath())) {
+            unlink($this->getResBundleFilePath());
+        }
+
         copy($this->getExtFilePath(), $this->getResultDir().$this->getExtFileName());
     }
 
     public function build($arg) {
+        $this->checkPreRequisites();
+
         $this->echoPaths();
 
-        $this->copyCode();
+        $this->prepareCode();
 
         $this->populateConfig();
 
@@ -95,4 +178,5 @@ class BuildOperaExtension extends BuildExtension {
 }
 
 $builder = new BuildOperaExtension();
+register_shutdown_function(array($builder, "cleanUp"));
 $builder->build(@$argv[1]);
